pyproject.toml
[project]
name = "legal-ai-expert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler[asyncio]>=3.11.2",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.1",
    "fpdf2>=2.8.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pillow>=12.1.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.22",
    "sse-starlette>=3.3.2",
    "google-auth>=2.48.0",
    "razorpay>=2.0.0",
    "langchain-anthropic>=1.3.4",
    "httpx>=0.28.1",
    "pandas>=3.0.1",
    "openpyxl>=3.1.5",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary]>=3.3.3",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "jinja2>=3.1.6",
    "slowapi>=0.1.9",
    "gunicorn>=25.1.0",
    "boto3>=1.42.73",
    "pyngrok>=7.5.1",
]
